+++ src/isla/language.py
@@ -2033,17 +2033,6 @@ class ExistsFormula(QuantifiedFormula):
         elif isinstance(new_in_variable, DerivationTree):
             new_in_variable = new_in_variable.substitute(subst_map)
 
-        new_inner_formula = self.inner_formula.substitute_expressions(subst_map)
-
-        if self.bound_variable not in new_inner_formula.free_variables() and (
-            self.bind_expression is None
-            or not any(
-                bv in new_inner_formula.free_variables()
-                for bv in self.bind_expression.bound_variables()
-            )
-        ):
-            return new_inner_formula
-
         return ExistsFormula(
             self.bound_variable,
             new_in_variable,